+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New curves classes: Hyperbola2D and Hyperbola3D.
 - Line: closest_point_on_line, from_point_and_vector
 - Line2D: get_slope, get_y_intersection.
+- New curves classes: Parabola2D/3D.
 
 ### Fixed
 - add missing name attributes to classmethods
@@ -25,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - 
 ### Unittests
 -Hyperbola2D/3D: line_intersections
+-Parabola2D/3D: line_intersections
 
 ## v0.13.0 [unreleased]